Move tx tests into a __tests_ dir
